@@ -1,7 +1,8 @@
 package cn.myproject.gorgeousgirl.Utils.glide4Lnterceptor;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.GlideBuilder;
