@@ -2,12 +2,13 @@ package cn.myproject.gorgeousgirl;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.util.Log;
 import android.widget.Toast;
+
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
@@ -20,8 +21,8 @@ import cn.myproject.gorgeousgirl.Utils.OKUtils;
 import cn.myproject.gorgeousgirl.Utils.StatusBarUtil;
 import cn.myproject.gorgeousgirl.data.DataBox;
 import cn.myproject.gorgeousgirl.eventBus.ItemImgEventBus;
-import cn.myproject.gorgeousgirl.recyclerview.LoadDataScrollController;
-import cn.myproject.gorgeousgirl.recyclerview.StaggeredAdapter;
+import cn.myproject.gorgeousgirl.view.recyclerview.LoadDataScrollController;
+import cn.myproject.gorgeousgirl.view.recyclerview.StaggeredAdapter;
 import cn.myproject.gorgeousgirl.view.ITestView;
 
 public class MainActivity extends Activity implements ITestView, LoadDataScrollController.OnRecycleRefreshListener {
@@ -31,6 +32,7 @@ public class MainActivity extends Activity implements ITestView, LoadDataScrollC
     private StaggeredAdapter mAdapter;
     private LoadDataScrollController mController;
     private OKUtils mOKUtils;
+    private final String URL = "https://www.mxnzp.com/api/image/girl/list/random";
 
     @Override
     protected void onStart() {
@@ -48,9 +50,7 @@ public class MainActivity extends Activity implements ITestView, LoadDataScrollC
     public void onItemGilde(ItemImgEventBus eventBus){
         if (eventBus != null){
             String url = eventBus.getUrl();
-            url = url.replace("http", "https");
-            url = url.replace("httpss", "https");
-//        final String newUrl = url;
+            if (!url.contains("https")) url = url.replace("http", "https");
 
             RequestOptions requestOptions = new RequestOptions().error(R.mipmap.ic_launcher);
             Glide.with(this)
@@ -74,8 +74,7 @@ public class MainActivity extends Activity implements ITestView, LoadDataScrollC
 //                Toast.makeText(MainActivity.this, "下拉", Toast.LENGTH_SHORT).show();
 //            }
 //        });
-
-        mRecyclerView = findViewById(R.id.recyclerview);
+        mRecyclerView = new RecyclerView(this);
         StaggeredGridLayoutManager layoutManager = new StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL);
 //        layoutManager.setGapStrategy(StaggeredGridLayoutManager.GAP_HANDLING_NONE);
         mRecyclerView.setLayoutManager(layoutManager);
@@ -87,15 +86,11 @@ public class MainActivity extends Activity implements ITestView, LoadDataScrollC
          * 设置监听
          */
         mRecyclerView.addOnScrollListener(mController);
+        mSwipeRefreshLayout.addView(mRecyclerView);
         mSwipeRefreshLayout.setOnRefreshListener(mController);
 
         mOKUtils = OKUtils.getInstance(this);
-        mOKUtils.get("https://www.mxnzp.com/api/image/girl/list/random");
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
+        mOKUtils.get(URL);
     }
 
     @Override
@@ -134,7 +129,7 @@ public class MainActivity extends Activity implements ITestView, LoadDataScrollC
     public void refresh() {
         //刷新的接口调
         DataBox.getmList().clear();
-        mOKUtils.get("https://www.mxnzp.com/api/image/girl/list/random");
+        mOKUtils.get(URL);
         //        设置数据加载结束的监听状态
         mController.setLoadDataStatus(false);
     }
@@ -144,7 +139,7 @@ public class MainActivity extends Activity implements ITestView, LoadDataScrollC
         //加载更多的接口回调
 //        pd.setMessage("正在刷新……");
 //        pd.show();
-        mOKUtils.get("https://www.mxnzp.com/api/image/girl/list/random");
+        mOKUtils.get(URL);
 //        设置数据加载结束的监听状态
         mController.setLoadDataStatus(false);
     }
